+++ contracts/interfaces/chainlink/AggregatorInterface.sol
@@ -2,17 +2,17 @@
 pragma solidity ^0.8.0;
 
 interface AggregatorInterface {
-    function latestAnswer() external view returns (int256);
+    function latestAnswer() external view returns (int);
 
-    function latestTimestamp() external view returns (uint256);
+    function latestTimestamp() external view returns (uint);
 
-    function latestRound() external view returns (uint256);
+    function latestRound() external view returns (uint);
 
-    function getAnswer(uint256 roundId) external view returns (int256);
+    function getAnswer(uint roundId) external view returns (int);
 
-    function getTimestamp(uint256 roundId) external view returns (uint256);
+    function getTimestamp(uint roundId) external view returns (uint);
 
-    event AnswerUpdated(int256 indexed current, uint256 indexed roundId, uint256 updatedAt);
+    event AnswerUpdated(int indexed current, uint indexed roundId, uint updatedAt);
 
-    event NewRound(uint256 indexed roundId, address indexed startedBy, uint256 startedAt);
+    event NewRound(uint indexed roundId, address indexed startedBy, uint startedAt);
 }
